requirements.txt

websocket-client
pandas==1.3.3
python-binance